pyproject.toml
[project]
name = "MicroC"
version = "0.1.0"
description = "Interpretador Micro-C"
readme = "README.md"
authors = [
    { name = "Lucas Chaves Itacaramby" }
]
requires-python = ">=3.10"
dependencies = [
    "ipdb>=0.13.13",
"lark-parser>=0.12.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-timeout>=2.4.0",
]

[project.scripts]
MicroC = "MicroC.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
